app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
databricks-sdk>=0.20.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
pyyaml>=6.0
